moveooc: Avoiding memleak when exiting
As we do malloc childdir, we need to be sure than everytime we do return
the function, this char is freed.

This patch consider that by default the return value is DEADBEEF and use
a simple goto jump to reach the end of the function when a return is
needed. This part of the code takes care of freeing childdir if required
and return the appropriate return code.